requirements.txt
# Kafka
kafka-python==2.0.2
python-snappy==0.6.1

# Spark
pyspark==3.4.0

# Databases
pymongo==4.6.1
psycopg2-binary==2.9.9
redis==5.0.1

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Data Generation
Faker==22.0.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Code Quality
black==23.12.1
flake8==7.0.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
